convert Q_rest to long when creating a congruence

+++ src/de/tilman_neumann/jml/factor/siqs/tdiv/TDiv_QS_1Large.java
@@ -232,7 +232,7 @@ public class TDiv_QS_1Large implements TDiv_QS {
 	
 		// Q is sufficiently smooth
 		if (DEBUG) LOG.debug("Sufficient smooth big factor = " + Q_rest);
-		return new Partial_1Large(A, smallFactors, Q_rest.intValue());
+		return new Partial_1Large(A, smallFactors, Q_rest.longValue());
 	}
 	
 	@Override
